rag-server/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
psycopg2-binary>=2.9.9
google-genai>=0.8.0
python-dotenv>=1.0.1
numpy>=1.26.0
redis>=5.0.0
